Relax rules when transitioning to provisioned
Relaxed the rules for stage_administrative_update() as per
recommendation in the Launchpad.

The provisioning state should have transitioned to 'provisioning' on
the host-unlock, it was observed to be still 'unprovisioned'.  This
commit makes the transition to PROVISIONED ('provisioned') less
restrictive, to also allow transition when constants.UNPROVISIONED
(not just the current constants.PROVISIONING) since the operational
state is being set to 'enabled’ (which can only be True if it is
unlocked-enabled, i.e. after provisioning).

The transition to PROVISIONING should've occured on the host-unlock
operation e.g., when following log occurs
"stage_administrative_update: provisioning"; which checks for whether
UNPROVISIONED or None there so this would align the OPERATIONAL_ENABLED
status with the PROVISIONED as well.

Closes-Bug: 1864874
Change-Id: Iaf093046dd3c315b6f22007e81b2b3f468f3e629
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>