Clear alarm after last device image deletion
The in-progress alarm is raised when the first device image is applied
to a device. If the host device image update is never issued and the
device images are deleted, then the alarm should be cleared as it is
returned to the initial state.
This update is to clear the alarm when the last device image is deleted.

Closes-Bug: 1883293

Change-Id: I113895731880ee5312bb915639df65b06430a421
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>