Upgrades: Add AIO partition data migration
In STX 4.0 the /scratch partition increased from 8G to 16G. This
increases the size of primary Linux LVM partition created by Anaconda as
compared to the size provided in STX 3.0.

To support the upgrade path from STX 3.0 to STX 4.0 we must migrate the
sysinv partition information for the primary Linux LVM partition and any
user created partitions on the root disk.

This migration allows the new partition layout on disk to match what is
expected in the sysinv database and will create any missing user defined
partitions on system unlock.

Change-Id: I060e7055d524d4ae44b595b9a172752aa5ac77ae
Closes-Bug: #1887192
Signed-off-by: Robert Church <robert.church@windriver.com>