Allow applying device image with non-device label
Since there are no FPGA devices on the system controller,
device label could not be assigned before the device image is applied.
This update allows applying device image with labels that are not
yet associated with any host device.

The applied attribute is also exposed in the device image API to indicate
whether the image was applied with or without labels.
An empty dict is displayed in the applied-label field if the device image
has been applied without labels.

The name-value pair parameter in device-image and device-label
CLI commands are changed to allow duplicate keys.

Added the cleanup of temporary device image file in the conductor.

Story: 2007875
Task: 40392

Change-Id: I84857dc216b9ca223560cd589b91704d65459d00
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>