Fix 20-exempt-admin-from-lockout.sh for IPv6
For IPv6 system, the script will fail with return status 1.
This is caused by curl fails to parse request url with IPv6 IP
address in "[ ]". The error produced by curl will be:

curl: (3) [globbing] error: bad range specification after pos 9

The fix is to add a "-g" option to the curl command.

Change-Id: I7f2c9c1eca81e845f2d443f3b8962ab206f1328d
Closes-Bug: 1887470
Signed-off-by: Andy Ning <andy.ning@windriver.com>