Fix run_docker_login DNS lookup to handle misconfigured nameservers
A customer added a non-existant nameserver and it resulted in an edge case
that broke the previous code.  (dig returned 9 even though it found an
AAAA record for the name.)

It looks like we're going to need to bite the bullet and explicitly issue
separate requests for IPv4 and IPv6.

Partial-Bug: 1886121
Change-Id: I8d653a82a2aee0a86af587f6392277974c9f1fef
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>