Update cgtsclient (system) to raise exception when auth_url is None
The cgtsclient is updated to check and raise an HTTPUnauthorized
exception when the auth_url is not set.  This can occur when the
session has expired or is invalidated and the auth_url is not set.

This is more meaningful exception than the 'TypeError' that was
being observed when the auth_url is None.

Change-Id: Id4249ca88ac4aeb102bfa82e631f22245a927a35
Story: 2007403
Task: 40451
Signed-off-by: John Kung <john.kung@windriver.com>