Fixes for device image alarm
This commit clears the in-progress alarm if there is no more
pending, failed or in-progress device images to be updated
after a device image is removed.

It also raises the alarm only when there are pending image updates.
If the system does not have any FPGA devices then the alarm is not
raised as there is no image update to be processed.

This commit also improved some comments about the applied field.

Story: 2007875
Task: 40392
Closes-Bug: 1883293

Change-Id: Ie8a8e6071ba9d5296992d46e410a7cbf19f8e1d1
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>