Fix a random tox linters failure in bashate
There was a file that was missing a newline which would
cause bashate to fail.
However the bashate command was being run as a list of inputs
rather than as individual files, so that error was being ignored
unless that bad file was the last file processed.
This submission fixes the file, and ensures that tox linters
job will fail is any file has an error, regardless of what is
processed afterwards.

Change-Id: I7587b7017b6855aad63dfaeaea5ecca0bcfc21ea
Closes-Bug: 1890287
Signed-off-by: albailey <Al.Bailey@windriver.com>