Fix label usage in device image commands
When an image is deleted, the device labels that are used to apply
the image and are not associated with any physical hardware are not removed.
The commit removes the device label if it is the last image associated
with the label and it is not associated with physical hardware.

The device image apply with multiple labels of the same device
for root-key image was incorrectly blocked. This update corrects
the validation of this case.

Closes-Bug: 1890299
Closes-Bug: 1890324
Closes-Bug: 1890340

Change-Id: Iaa34b9fe57678d6283a4cce4a30d7df4673ec996
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>