Add API to import load metadata for SX upgrade
For SX upgrades we can skip importing the entire bootimage ISO before
starting the upgrade. With duplex systems the ISO data is used to
install the N+1 load on other hosts. For SX systems there are no other
hosts, so transferring the ISO is not needed. For now this feature will
be restricted to SX subclouds.

When the metadata import is used patching will not be available for that
load. If patches are required for the N+1 load they will need to be
applied either via a patched ISO or with sw-patch before running the
upgrade_platform playbook.

Change-Id: I898786de42798191f9ec7b14d942456651f813a5
Story: 2007403
Task: 40806
Signed-off-by: David Sullivan <david.sullivan@windriver.com>