Update drbd resize operation to run after puppet manifest
There is a potential race condition in triggering the filesystem
resize check.  The resize check is triggered on the start of
a controller filesystem drbd resize operation, however, if another
configuration is applied before the resize check may be early,
in which case the lv would be resized but not drbd.

In the drbd manifest update to trigger the drbd filesystem resize
check after the lv filesystem  have been resized.

Change-Id: Idd0629f77c414a79cc1969ff1110d5dbd3548cda
Closes-Bug: 1894003
Depends-On: https://review.opendev.org/#/c/749596/
Signed-off-by: John Kung <john.kung@windriver.com>