Ensure admin endpoint cert always applies to both controllers
Previous commit incorrectly skip applying admin endpoint cert when
the cert is identical to the one exists on active controller. This
causes the certificate is not applied to controller-1 when it is
unlocked the 1st time, and potential some other cases to the standby
controller.

The change ensure the admin endpoint certificate is always applied.

Story: 2007347
Task: 40624

Change-Id: Idfb0063f765fbcde70bbd116ff953564180dcd98
Signed-off-by: Bin Qian <bin.qian@windriver.com>