Add timeouts for helm commands
Armada is still connecting tiller to manage releases. To make it
more robust, we should set "--tiller-connection-timeout" to the
helmv2 commands to prevent helm commands hang in case tiller is
not running or having connection issue. An additional operation
timeout is added for some commands as well to enhance the robustness.

Change-Id: I5b7bb06fec254455602a183e9151a8abe8541517
Closes-Bug: 1896529
Signed-off-by: Angie Wang <angie.wang@windriver.com>