Use stx_admin account for flock service to communicate with openstack
admin account is used before, but if admin password is changed, flock
service cannot be notified and cannot get the new password, so flock
service like nfv-vim cannot fetch openstack vm info ever.
stx_admin account is created for this case. It is created by openstack
app when it is applied, and password is saved to keyring so flock
service could use it to communicate with openstack app.
helm is updated to use stx_admin account instead of admin account.

Partial-Bug: 1887755

Change-Id: I8b7ce7b60ccb3f0d0a595395b591565b24a590d5
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>