Controller-1 fails to unlock after data-migration
During an upgrade, data-migration will complete, but the inv_state for
controller-1 will remain at reinstalling. Consequently it will not be
possible to unlock controller-1.

This is addressed by starting the sysinv-agent after the data migration
is complete. The sysinv-agent will then report the inv_state to
controller-0.

Story: 2008055
Task: 40971
Change-Id: I96eb5359bc6b7438c63f1af376db5d8a4a7327f5
Signed-off-by: David Sullivan <david.sullivan@windriver.com>