Preserve quotations when loading yaml using ruamel.yaml
RoundTripLoader in ruamel.yaml will omit superfluous quotation
marks. In some cases, the result by loading a yaml with and
dumping it out will be unreadable by PyYAML. For example, the
pattern:
  hosts: ["mon-logstash:port"]
will be transformed to:
  hosts: [mon-logstash:port]
Currently, the Armada project is using PyYAML. It's better to keep
the  quotes to allow the following process have to use PyYAML.

This commit adds the parameter preserve_quotes=True to the
RoundTripLoader to perserve the quotes when loading the original
charts.

Tested with a fresh install, upload the stx-monitor tarball with right
formatted yaml files and successful apply the application.

Change-Id: I1fd16b24bfb62ffc3bc885babbf6eb8a9356c421
Depends-on: https://review.opendev.org/#/c/754501/
Partial-Bug: 1896530
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>