Migrate to database backend for backup and restore
Drop the use of a flag file for restore.
Create the database table for backup and restore.
Use the database to store information.

CREATE TABLE backup_restore (
        created_at TIMESTAMP,
        updated_at TIMESTAMP,
        deleted_at TIMESTAMP,
        id serial PRIMARY KEY,
        uuid VARCHAR ( 36 ) UNIQUE NOT NULL,
        state VARCHAR ( 128 ) NOT NULL,
        capabilities TEXT
);

This is an improvement to allow backup and restore functionality to be
extended in an easier manner. Not required to fix the the bug.

Added unit tests.

Depends-On: I7b7fab99d457056032dbbd612363cd5036736cda
Depends-On: I44fc4aaa528e372a84115714f271b4f5e063f86e
Partial-Bug: 1887648
Change-Id: Ibb96696a35fe7b560aa002c442e2af735d08ec24
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>