Split ip address without port
The url string should split by slash to get the ip address in the
case there is no specified port. This commit splits the path and
port from the ip address before validating, allows the ip validate
methods get the correct format of address. After this change, the
port sector in the is_valid_domain is no longer required.

An unit test against the is_valid_domain_or_ip method with a
list of typical urls is added.

Tested wih a fresh install on an AIOSX setup, and boot it up with
the ansible playbook.

Closes-bug: 1898921
Change-Id: I5d37fafbfd7bdc669bba61e5d0c04eb467c46e70
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>