Add support for disk wiping without GPT formatting
To support a Rook based Ceph cluster, provide an option when wiping
disks to skip adding a partition table. This allows preparing disks on
all hosts during provisioning from the active controller that will be
formatted and run as OSDs when the Rook k8s application is applied.

Change-Id: I93a9770b0d78ddddc01fe7956d7ad7058acc8e71
Story: 2005527
Task: 41127
Signed-off-by: Robert Church <robert.church@windriver.com>