Use kernel module as port virtual function driver
For network interfaces, the supported kernel module should be
used as the virtual function (VF) driver for the underlying
port.  This allows a determination of the driver to use when
one selects an interface VF driver of 'netdevice'

This commit fixes an issue that can occur if the parent
interface in an SR-IOV 'split-nic' scenario has a driver of
vfio rather than netdevice.  In that case, the port's VF
driver was erroneously set to vfio-pci, which would be used
instead of the supported kernel device for the set of netdevice
VFs on the interface.

Change-Id: Ie863e32d04f9636d52a65ce4d56c441f6b69e3a9
Closes-Bug: 1885446
Signed-off-by: Steven Webster <steven.webster@windriver.com>