add sysinv support for specifying cpu function by range
In order to add flexibility we want to allow specifying CPU
function by range, rather than just by count.

This will allow us to run something like this on the CLI:

system host-cpu-modify -f application-isolated -c 3-5,25 controller-0

There are a couple complications to be aware of.  First, sysinv will
NOT automatically add any missing SMT hyperthreads if the host has
hyperthreading enabled.  Second, when specifying CPU function for
a different function via the CLI the range specification is lost and
gets converted to a simple count.  This implies that (for the CLI)
only one function can support a range-based specification, and it
must be specified last.

Story: 2008760
Task: 42180
Change-Id: Id21d9968b6b0b59e163f42098be7a6f0e6ef739d
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>