Report port and device inventory after the worker manifest
Normally, the SR-IOV configuration of a network or device is retained
across reboots.  In most cases, this is sufficient for the
sysinv-agent to report port and device inventory at any time after it
is started.

A problem can occur however for the case of an N3000 FPGA device.
This device requires a reset on every reboot, which clears the SR-IOV
configuration until the puppet worker manifest has completed the
(re)configuration.  In this case, there is a small chance that the
sysinv-agent audit (every 60 seconds) will run in-between the reset
and the driver configuration.  Since the agent will only actually
report the port and device inventory once, a problem can occur
after a second host lock-unlock, as the SR-IOV configuration data is
not accurately reflected in the db.

This commit follows a similar method used by the sysinv agent's
hugepage reporting, in that the port and device configuration are
only reported after the worker manifest has completed.

Change-Id: Id4af97e0175482d561745a5cb34650dc6b5653b8
Closes-Bug: 1901968
Signed-off-by: Steven Webster <steven.webster@windriver.com>