Add variables for snmp in fm.conf
Snmp trap client needs the following three variables
to connect to snmp trap server.
- trap_server_ip
- trap_server_port
- snmp_enabled
This commit is to update sysinv puppet hieradata with
the value of snmp_enabled based on whether the snmp
app is applied.
trap_server_ip and trap_server_port are fixed.
snmp_enabled takes 1(True)/0(False) depends on snmp
armada app is applied or not (1 when applied).

Change-Id: I76f9ef3e9ca6a98c0cd37a72f7aaccd10353b26b
Story: 2008132
Task: 41205
Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>