AIO upgrades: update partition data migration script
The size of primary Linux LVM partition for cgts-vg created
by Anaconda in stx5.0 for AIO is reduced compared to the
size in stx4.0.

To support the upgrade from stx4.0 to stx5.0, we need to
migrate the sysinv partition information for the primary LVM
partition by creating a new partition entry and assigning
it as a physical volume to the cgts-vg to maintain the stx4.0
size and adjust any user created partitions on root disk.

Update the existing migration script to support that. The new
partition and user defined partitions will be created on unlock.

Tested the AIO-SX and AIO-DX upgrade with user created two
partitions(cgts-vg, nova-local) on root disk and user created
partitions on second disk.

Closes-Bug: 1892554
Change-Id: I8e6e4246e978b0d2e87626f6c980d4e56ff0f56f
Signed-off-by: Angie Wang <angie.wang@windriver.com>