Merges key/value pairs in platform.conf
Change migrate_platform_conf method inserts multiple copies of many
key value pairs from the backup file to the new installed file which
results duplicated values for a certain key. This commit
updates/preserves the value for the keys to insure every key has a
unique value.

Tested with building a new iso and upgrading an AIOSX with the new
iso, the upgrade activated and completed successfully. In the
/etc/platform/platform.conf, every key has a unique value. The values
of the keys listed in the "skip_options" list equals to the previous
values before upgrading.

Closes-bug: 1891935
Depends-on: https://review.opendev.org/#/c/762554/
Change-Id: I250f49d8ce4f2d06d068017c46ac8bd0c08b69be
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>