Fix live-migration for boot-from-volume VM
The reason why live-migration is failing on AIO-DX
is because cinder-volume is hanging while running
'ceph mon dump' command.

'ceph mon dump' command reads ceph.conf file to get the
addresses of all the monitors. It leads to delays
if those addresses don't point to valid monitors.

On AIO-DX setup, ceph.conf had 3 monitors and this
caused delays of up to 1 minute when 'ceph mon dump'
was executed inside cinder pod.

To get the correct response time for 'ceph mon dump'
in the cinder pod, a single monitor pointing to the floating
IP should be used for AIO-DX.

Closes-bug: 1904594
Change-Id: I70ecd5c468383358f7566061b65bf11917fe904a
Signed-off-by: Elena Taivan <elena.taivan@windriver.com>