Add host command support for the edgeworker node
This is an experimental feature in stx5.0.

This commit enables following commands for the edgeworker node:
  system host-add/system host-update/system host-delete
After the host being added/updated, the mgmt ip of an edgeworker
node will be assigned during the configuration process of it.

There will be limitations of edgeworker nodes before the final
phase of the feature finished:
- The Kubernetes provisioning requires ansible playbook triggered
  manually.
- Gather node HW information is not supported.
- Configure node from controller is not supported.
- Manage node lifecycle is not supported.
- Update/upgrade node is not supported.

Story: 2008129
Task: 40862

Change-Id: I7e6de65ba848d9468a4e5afddd16b1cd9e3cd7dd
Signed-off-by: Mingyuan Qi <mingyuan.qi@intel.com>
Depends-On: https://review.opendev.org/c/starlingx/config/+/761716