Fix the exception of host label assign
If Kubernetes node has images without tag and digest
("crictl images" shows <none>:<none>), system operations
like "system host-label-assign" that sending the requests
to Kubernetes to patch/query node via python Kubernetes
client will raise this exception.
  "Invalid value for `names`, must not be `None`"

The exception is raised in python Kubernetes client but not
from kubernetes. Host label is actually created successfully
in Kubernetes.

Implement a workaround to replace V1ContainerImage.names
in the python Kubernetes in the python Kubernete client
to bypass the "none image" check.

This workaround should be removed if the proposed solutions
can be made in kubernetes or a workaround can be implemented
in containerd.
https://github.com/kubernetes/kubernetes/pull/79018
https://github.com/containerd/containerd/issues/4771

Change-Id: I2baae31b54c364c7d908f87b32ca0181a032976c
Closes-Bug: 1905481
Signed-off-by: Angie Wang <angie.wang@windriver.com>