Wait for all apps are stable before activate
During an upgrade on an AIO-SX, images are automatically recovered after
migrating to the new release and the host is unlocked. While this is
occurring, applications may be in an 'applying' or 'restore-requested'
state. The recovering period may time consuming if the network(mgmt) is
not stable or in low speed. This commit adds retries up to 15 mins to
wait for all the apps are uploaded or applied for AIO-SX system.

Tested with upgrading and activate an AIODX central cloud and an
AIO-SX subcloud. In the orchestrator, if an app on subcloud is still
in applying status for more than 15 mins, the activate-audit fails
once, the orchestration succeeds after the app status becomes applied.

Partial-Bug: 1904934
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>
Change-Id: Ia995b7d6785dfdd0420e977b4431f68ac3b495c3