Fix tox failures from new version of virtualenv and pip
New pip was released on Dec 3, 2020 which breaks downloads
from pypi for some URLs (python 3).

New virtualenv was released Dec 7 which includes that new pip.

In order for sysinv in python3 on ubuntu-bionic to pass its
tox and zuul jobs, the legacy resolver must be used.

The requirements may need to be updated in future revisions.

Closes-Bug: 1907125
Change-Id: Idecaee7d83f4cbf02454e172aa254b7d8d449141
Signed-off-by: albailey <Al.Bailey@windriver.com>