Consistent creation of registry secret
In this commit, the default-registry-key is created from sysinv
credentials during the apply of the very first platform app
(nginx-ingress-controller) in kube-system namespace. This key
can no longer be removed. It will be used to create registry secret
for all helm-based applications in their namespaces upon
applying the application.

Tests in AIOSX:
  - Install the system.
  - Remove cert-manager app, prune containerd image cache to force
    repull of the images from the registry, reapply the app.
  - Remove platform-integ-apps, verify that the default-registry-key
    in kube-system ns is not removed, reapply platform-integ-apps
  - Apply oidc-auth-apps, remove the app, verify that the
    default-registry-key in kube-system ns is not removed.

Closes-Bug: 1906337
Change-Id: I042b1bf81f44a8c5498661c0bfca219c7150c57a
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>