API documentation error for partitions
An error occurred when using the partitions part
of the API documentation
(POST /v1/ihosts/​{host_id}​/partitions). After analyzing the
code, I found that the URLs set in the code are /v1/partitions.
This is inconsistent with the API documentation,
so update the partitions section of the API documentation.

Closes-Bug: #1897231

Change-Id: I9fc61b86bdfc4cb12acfd4f4cd8ecf8853ab338e
Signed-off-by: YuehuiLei<leiyuehui@inspur.com>