API documentation error for Ceph Storage Functions
An error occurred when using the Ceph Storage Functions part
of the API documentation (POST  /v1/ihosts/​{host_id}​/istors).
After analyzing the code, I found that the URLs set
 in the code are /v1/istors.
This is inconsistent with the API documentation,
so update the partitions
section of the API documentation.
And I found that the request parameters are mandatory
rather than optional.

Closes-Bug: #1897252
Change-Id: I672735a9d1fa30894dfc0a9b2c12888a1e6b6271
Signed-off-by: YuehuiLei<leiyuehui@inspur.com>