Update expired unit test certificate
The sysinv unit test certificate expired Dec 22, 2020
This extends the date to Dec 22 2021.

The unit test header was updated to trigger zuul to verify
the changes are valid.

A Readme file has been provided which lists the expiry dates
for all the pem files used for the unit tests. Information
on how to update the pem files will be added to the readme
when all the expiry dates are aligned.

Change-Id: I63cdd865a1818756f9ec3f20ebd724d2ad2d5729
Partial-Bug: 1909817
Co-Authored-By: Sabeel Ansari <Sabeel.Ansari@windriver.com>
Signed-off-by: albailey <Al.Bailey@windriver.com>
(cherry picked from commit 97ad36382ae7abeba0cf1118f3e80683f2d6215a)