sysinv: Intel ACC100 (Mt Bryce) enablement
This commit adds SR-IOV device plugin support for forward error
correction (FEC) devices that are enabled on an Intel ACC100 (Mt.
Bryce). The Intel ACC100 is mounted on Lisbon ACC100 Card.

The FEC device is intended for use by a DPDK application. It is
presented to the system under resource name: "intel_acc100_fec."

An example usage to modify the device:

system host-device-modify <host> <device_name> \
  -e true \
  --driver igb_uio \
  --vf-driver <driver> \
  -N <num_vfs>

And example assignment to a pod:

 resources:
      requests:
        memory: 4Gi
        intel.com/intel_acc100_fec: '16'
        cpu: 6
      limits:
        hugepages-1Gi: 2Gi
        memory: 4Gi
        intel.com/intel_acc100_fec: '16'
        cpu: 6

Story: 2008440
Task: 41403

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/775253
Depends-On: https://review.opendev.org/c/starlingx/integ/+/775252

Signed-off-by: Babak Sarashki <Babak.SarAshki@windriver.com>
Change-Id: I831fd16a0410ee988365c067789f760139274ec8