Add api call to trigger creation of gnp
Creating an oam network was not correctly setting the flag to trigger
the runtime puppet manifest to create the default globalnetwork policy
and hostendpoint resources.

This change adds the correct api call to ensure that the flag is set
when an oam network is created. This ensures that the default
globalnetworkpolicy and hostendpoint kubernetes resources are created as
intended.

This issue was previously masked by the fact that the relevant runtime
manifest was being run on every single unlock. This behaviour was
changed and revealed that the required flag was not being set upon oam
creation.

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/1911213

Signed-off-by: Cole Walker <cole.walker@windriver.com>
Change-Id: I5b570e82246d82a4243ee68a871333876d231c85