Ensure Armada lock is released
If the apply/remove of an application is failed due to
an exception of the Armada request or an abnormal exit
of the Armada request, the Armada lock cannot be released
by Armada which causes the subsequent re-apply of the
application to fail as it cannot acquire the lock.

Update to delete the lock if an exception occurs during
Armada request or Armada request exits abnormally.

Closes-Bug: 1911243
Change-Id: I6ab02474831152b5b2b7302f0799c48d05fef64d
Signed-off-by: Angie Wang <angie.wang@windriver.com>