Migrate etcd after both controllers are upgraded
The etcd database can go out of sync if we swact back and forth between
controllers running different versions. This can happen if we need to
abort the upgrade after swacting to controller-1. Since we are not
upversioning etcd in this release we will address this by waiting to
migrate the etcd data when both controllers are running the new release.
The migration will now take place during the swact to controller-0
before upgrade-activate.

This solution will present some problems when we do upgrade etcd, so
further development will be required at that time.

Story: 2008055
Task: 41630
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Change-Id: I02b82bfe1a4b4b69aaa85d5f0d20246b9cda5629