Update config tracking for reboot required config
The reboot config tracking needs to be updated to ensure
it aligns the target with the actual applied config.
It must also initialize and only clear the config for
the active controller if a reboot has actually occurred.

On host-swact, the target controller is checked for config
up to date condition, rather than both source and target
controller; as the source active controller would still
tracked the config requirement in the persisted database.

Improve traceback of config requests.

Tests Performed:
  Perform host-swact with active controller config out of date
  Perform host-swact with standby controller config out of date
  Verify AIO-SX and Duplex config operations
  Restart sysinv with reboot config set and perform runtime config
  Perform runtime config while reboot required config set
  Perform runtime config while reboot required config cleared

Change-Id: I339ad82a2c7b37ac1c97c3eb790a231a40914250
Closes-Bug: 1914085
Signed-off-by: John Kung <john.kung@windriver.com>