Fix out-of-date alarm wrongful removal
When updating docker service parameters by either adding or
deleting them, a "Configuration is out-of-date" alarm will
automatically appear for both active and standby nodes.
The alarms should only be cleared by a manual lock/unlock
of each node.
The fix addresses the scenario where a configuration update
that triggers a runtime manifest clears the existing out-of-date
alarm (e.g.: system certificate-install -m ssl_ca <cert>).
The issue is reproduced no matter what configuration update with
runtime manifest is used or if the existing alarm was due to a
“service-parameter-add” or a “service-parameter-delete” command.

Closes-Bug: 1878018

Signed-off-by: Carmen Rata <carmen.rata@windriver.com>
Change-Id: I1e545327ae394385c995aee885590779432c60b0