Add support for AIO-SX to DX migration on subcloud
This update allows the user to change the system_mode on a subcloud
from simplex to duplex using the system modify command. The sysinv DB
and the platform.conf will be updated with the new system mode. The
semantic checks were modified to ensure that changing from duplex to
simplex is prohibited. The changes also include support for updating
the OAM networking config using the oam-modify command.

Story: 2008587
Task: 41885

Signed-off-by: Melissa Wang <melissa.wang@windriver.com>
Change-Id: If7c14222ca66323225400ed88f214655f33fe615