Remove unsafe umask usage
The sysinv code runs under eventlet that causes the
running greenthread to swap out the original umask. This
results in the sysinv code running with the incorrect umask.

This can be demonstrated by the "system dns-modify" command,
the agent process starts with a umask of 022, switched to 0,
and is never restored.

This simple fix is to audit where os.umask is being used and
replace os.umask with os.chmod.

Testing has been done locally by:

1. Running the "system dns-modify nameservers=8.8.8.8,8.8.4.4" command
   and checking the results/permissions of /etc/resolv.conf. Also,
   cheked to see if the umask flag have been changed in /proc/XXX/status
   before and after running the command.
2. Running an "system applicaton-upload" command on an installed helm
   armada package, these are located in /usr/local/share/application/helm.
   After the application upload, the application-apply should be
   "applied" without error/failure as shown in "system
   application-list".
3. Running a distributed-cloud and checking for any errors. The command
   "dcmanager subcloud show <subcloudname>" should show the identity
   service in sync after the dcmanager subcloud manage <>".

Closes-Bug: 1915955

Signed-off-by: Charles Short <charles.short@windriver.com>
Change-Id: I16ce695cfc4f6fb496ac0b3287906cc968ec5e98