Fixes the time calculation in the is_expired method of the Token class
The previous implementation calculated an absolute delta between the
expiration time and the current time and compared it to a time window.
The calculation should be done without getting the absolute value
to account for an expiration time older than the time window.

Added unit tests for the affected code.

Closes-Bug: 1915952
Signed-off-by: Isac Souza <IsacSacchi.Souza@windriver.com>
Change-Id: I46bb36993d6f02978a593a1c1d49692e627d2a9b