Support background runtime manifests during upgrade-activate
In distributed cloud environments runtime manifests can be applied in
the background. This can cause hosts to become config out-of-date after
the upgrade-activate completes. This is due to the large window between
setting the host's config_target and updating the config_applied. If a
manifest is run in this window the host will remain config out-of-date
until a lock/unlock is performed.

To address this the config_target changes will be limited to hosts that
apply a runtime manifest as part of the upgrade-activate process.
Further the config_target will be updated immediately before the
_config_apply_runtime_manifest is called.

Story: 2008055
Task: 41917
Change-Id: I2e60c7557e8d398eeef2a407a0552f5e8f4a1f18
Signed-off-by: David Sullivan <david.sullivan@windriver.com>