Progress adjust metadata refactoring
Changed the name of the constant and the yaml key to better reflect
the purpose. Now the value is an integer which represents the
adjustment value used to compute the percentage completion when
applying charts. Cleaned up the code around the usage of the value
and computing the percentage.

Story: 2007960
Task: 41959

Signed-off-by: Andrei Grosu <andrei.grosu@windriver.com>
Change-Id: Ia3b07b83762cdf20f6809222dc687f67c15deee5