Allow modification of OAM IP addr, in AIO-SX, without locking the host
Users will be able to change the OAM IP address without a lock/unlock
cycle. To achieve this some services will be restarted (sm, sm-api,
haproxy and vim-webserver) to reopen the L4 ports using the old IP
address as part of the socket.

Some config files in /etc are being updated also with the new address.

Story: 2008531
Task: 42060
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/780955
Change-Id: I9e77fc60882f20d4f31c3e38b5305b1f207f40d9
Signed-off-by: Andre Fernando Zanella Kantek <AndreFernandoZanella.Kantek@windriver.com>