Restart tiller on openstack pending install check
When the armada-api pod, which runs helmv2-cli, goes up, it connects
tiller to a postgres instance running on the active controller. It does
that using the active controller's floating IP address. On setups with
more than one controller, this creates an issue where that connection
is no longer valid after performing a host swact, since it still points
to the old controller.

After the swact, the first helmv2-cli command will fail with a broken
pipe error. One of the steps in the Openstack installation involves
checking the system for pending helm installations, which will run a
helmv2-cli command, causing the application operation to fail if ran
right after a host swact.

This patch is a workaround to that problem. It forces a tiller restart
by catching the first HelmTillerFailure exception caused by the broken
pipe error and retrying the operation, which then will reestablish the
connection between tiller and the correct instance of postgres.

Closes-Bug: #1917308
Change-Id: Ia09f9d2844611471314d5d3af70e9bbb0938437c
Signed-off-by: Gustavo Santos <gustavofaganello.santos@windriver.com>