sysinv-api OCF script API request
In order to avoid other services requesting to sysinv-api before it is
ready to handle request'; and to avoid individual services to
implement their own retry logic, sysinv-api OCF script was changed
to verify if sysinv-api is providing service before returning it is
ready.

Testing:
1. Bootstrap of AIO-DX
2. host-swact on AIO-DX
3. Upgrade path on AIO-DX, 20.06 to 21.05 (including host-swact)
4. subcloud bootstrap
5. Double sysinv-api kill causing swact

In all tests, the logs are verified to confirm the retry logic is
engaged and sysinv-api is properly started, also cert-mon.

Closes-Bug: 1913455
Signed-off-by: Adriano Oliveira <adriano.oliveira@windriver.com>
Change-Id: Ia17e9f7a15602c0cc52cb01896fac42ce4fcdcb9