Run SRIOV manifest during controller-1 migration
To address another issue we blocked manifests from applying on a
controller with a different version from the active controller. This
blocks the sriov manifest from running on controller-1 after it has been
upgraded to N+1. To address this the data migration will check if
controller-1 has sriov interfaces configured and apply the manifest if
necessary.

The previous change:
https://review.opendev.org/c/starlingx/config/+/766637

Closes-Bug: 1921788
Change-Id: I1ec790fced00a4e973e546a260d28f52ef06fb3a
Signed-off-by: David Sullivan <david.sullivan@windriver.com>