Fix resize of drbd filesystems
This commit adds the following modifications:

- The drbd filesystems sizes are now calculated using 'dumpe2fs'
utilitary because it gives better results for larger filesystems.

- Added extra checks before and after executing 'resize2fs'.
Before running 'resize2fs' check if the drbd device is resized using
"/sys/block/{drbd_device}/size" and sector size.
After running 'resize2fs', check if the filesystem is resized
using 'dumpe2fs'.

- The drbd filesystems were resized only if they were
in SyncSource or PausedSyncS states. There are cases
when drbd-overview showed “Connected” instead of this sync states
and the filesystems would never be resized. Now
the drbd filesystems will also be resized when they
are in "Connected" state.

Closes-Bug: 1921896
Change-Id: I548300deb8916ce863bcd4bb70969cb9d51c9c2a
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>