Python 2 to Python 3 compatibility
The changes of https://review.opendev.org/c/starlingx/config/+/782575
was used to test the cgts-client, so it is need to be merged first.
Removing python-neutronclient because this dependency is unnecessary,
it was removed by copying a few very small utility functions from
python-neutronclient into the cgtsclient.

Development: When I was trying to find things to modify I followed the
approach of build the client, get the tar file, I set up 2 environments
one based on python2 and another python3, I installed the tar client
in both environments and i exported the env vars that the client expect
to get to request the controller, and doing that I could switch between
the two python and indentifying what I should modify.

Test: After all the modification I built an ISO and I installed that
to run some commands and check if my changes got any side effects. After
that followed the procedure to update the remote CLI docker image and
insert the updated client there and I test this new image in the
remote CLI.

Story: 2007106
Task: 42268

Depends-On: I5086832605752bdb00a40a24596494c8fd987692
Signed-off-by: Rafael Jardim <rafaeljordao.jardim@windriver.com>
Change-Id: Ibf919260693f1cbe99993d1de01ecf785d604839